+++ Projects/Printf/ft_print_pointer.c
@@ -6,13 +6,13 @@
 /*   By: pdel-olm <pdel-olm@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/24 00:10:22 by pdel-olm          #+#    #+#             */
-/*   Updated: 2024/02/28 02:08:23 by pdel-olm         ###   ########.fr       */
+/*   Updated: 2024/02/28 03:08:11 by pdel-olm         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_print_pointer(long p)
+int	ft_print_pointer(size_t p)
 {
 	int	count;
